refactor(apps/web/인게임): CanvasLoader 컴포넌트 삭제하고 기능을 Canvas 컴포넌트로 이전
